improved rendering
