+++ packages/web/src/app/(app)/@sidebar/(routes)/(default)/page.tsx
@@ -1,0 +1,5 @@
+import { DefaultSidebar } from "../../components/defaultSidebar";
+
+export default async function Page() {
+    return <DefaultSidebar />;
+}
