+++ packages/web/src/app/(app)/@sidebar/(routes)/settings/[...slug]/page.tsx
@@ -1,0 +1,5 @@
+import { SettingsSidebar } from "../../../components/settingsSidebar";
+
+export default async function Page() {
+    return <SettingsSidebar />;
+}
