+++ packages/web/src/app/(app)/layout.tsx
@@ -19,10 +19,9 @@ import { notFound, redirect } from "next/navigation";
 import { PendingApprovalCard } from "./components/pendingApproval";
 import { SubmitJoinRequest } from "./components/submitJoinRequest";
 import { hasEntitlement } from "@sourcebot/shared";
-import { OrgRole } from "@sourcebot/db";
 import { env } from "@sourcebot/shared";
 import { GcpIapAuth } from "./components/gcpIapAuth";
-import { getAnonymousAccessStatus, getConnectionStats, getMemberApprovalRequired, getOrgAccountRequests } from "@/actions";
+import { getAnonymousAccessStatus, getMemberApprovalRequired } from "@/actions";
 import { JoinOrganizationCard } from "@/app/components/joinOrganizationCard";
 import { LogoutEscapeHatch } from "@/app/components/logoutEscapeHatch";
 import { GitHubStarToast } from "./components/githubStarToast";
@@ -33,17 +32,16 @@ import { getPermissionSyncStatus } from "../api/(server)/ee/permissionSyncStatus
 import { ServiceErrorException } from "@/lib/serviceError";
 import { ConnectAccountsCard } from "@/ee/features/sso/components/connectAccountsCard";
 import { SidebarProvider } from "@/components/ui/sidebar";
-import { AppSidebar } from "./components/appSidebar";
-import { SidebarOverrideProvider } from "./components/appSidebar/sidebarOverrideContext";
-import { getUserChatHistory } from "@/features/chat/actions";
 
 interface LayoutProps {
-    children: React.ReactNode,
+    children: React.ReactNode;
+    sidebar: React.ReactNode;
 }
 
 export default async function Layout(props: LayoutProps) {
     const {
-        children
+        children,
+        sidebar,
     } = props;
 
     const org = await __unsafePrisma.org.findUnique({
@@ -168,30 +166,6 @@ export default async function Layout(props: LayoutProps) {
     const isPermissionSyncBannerVisible = session && hasEntitlement("permission-syncing");
     const hasPendingFirstSync = isPermissionSyncBannerVisible ? (await getPermissionSyncStatus()) : null;
 
-    const chatHistory = session ? await getUserChatHistory() : [];
-    if (isServiceError(chatHistory)) {
-        throw new ServiceErrorException(chatHistory);
-    }
-
-    // Compute sidebar notification flags
-    const isSettingsNotificationVisible = await (async () => {
-        if (!session) {
-            return false;
-        }
-        const membership = await __unsafePrisma.userToOrg.findUnique({
-            where: { orgId_userId: { orgId: org.id, userId: session.user.id } },
-            select: { role: true },
-        });
-        if (membership?.role !== OrgRole.OWNER) {
-            return false;
-        }
-        const connectionStats = await getConnectionStats();
-        const joinRequests = await getOrgAccountRequests();
-        const hasConnectionNotification = !isServiceError(connectionStats) && connectionStats.numberOfConnectionsWithFirstTimeSyncJobsInProgress > 0;
-        const hasJoinRequestNotification = !isServiceError(joinRequests) && joinRequests.length > 0;
-        return hasConnectionNotification || hasJoinRequestNotification;
-    })();
-
     return (
         <SyntaxGuideProvider>
             {
@@ -205,22 +179,16 @@ export default async function Layout(props: LayoutProps) {
                 ) : null
             }
             <div className="fixed inset-0 flex bg-shell">
-                <SidebarOverrideProvider>
-                    <SidebarProvider defaultOpen={cookieStore.get("sidebar_state")?.value !== "false"}>
-                        <AppSidebar
-                            session={session}
-                            chatHistory={chatHistory}
-                            isSettingsNotificationVisible={isSettingsNotificationVisible}
-                        />
-                        <div className="flex-1 min-h-0 flex flex-col pt-2 pb-2 pr-2">
-                            <div className="flex-1 min-h-0 bg-background flex flex-col border border-[#1d1d1f] rounded-xl">
-                                <div className="flex-1 min-h-0 overflow-y-auto">
-                                    {children}
-                                </div>
+                <SidebarProvider defaultOpen={cookieStore.get("sidebar_state")?.value !== "false"}>
+                    {sidebar}
+                    <div className="flex-1 min-h-0 flex flex-col pt-2 pb-2 pr-2">
+                        <div className="flex-1 min-h-0 bg-background flex flex-col border border-[#1d1d1f] rounded-xl">
+                            <div className="flex-1 min-h-0 overflow-y-auto">
+                                {children}
                             </div>
                         </div>
-                    </SidebarProvider>
-                </SidebarOverrideProvider>
+                    </div>
+                </SidebarProvider>
             </div>
             <SyntaxReferenceGuide />
             <GitHubStarToast />
